[Refactor] places-> place 파일명으로 변경되면서 Import 변경
